Enabling PCA transform algorithm support for GPU (#250)
Enabling PCA transform algorithm support for GPU

+++ service/kernel/oneapi/blas_gpu.cpp
@@ -63,37 +63,38 @@ services::Status ReferenceGemm<algorithmFPType>::operator()(const Transpose tran
     args.set(9, (uint32_t)offsetB);
     args.set(10, beta);
     args.set(11, c_buffer, AccessModeIds::write);
-    args.set(12, (uint32_t)one);
+    args.set(12, one);
     args.set(13, (uint32_t)ldc);
     args.set(14, (uint32_t)offsetC);
 
     if (transa == Transpose::NoTrans && transb == Transpose::NoTrans)
     {
         args.set(3, one);
-        args.set(4, lda);
+        args.set(4, (uint32_t)lda);
         args.set(7, one);
-        args.set(8, ldb);
+        args.set(8, (uint32_t)ldb);
+
     }
     else if (transa == Transpose::Trans && transb == Transpose::NoTrans)
     {
-        args.set(3, lda);
+        args.set(3, (uint32_t)lda);
         args.set(4, one);
         args.set(7, one);
-        args.set(8, ldb);
+        args.set(8, (uint32_t)ldb);
     }
     else if (transa == Transpose::NoTrans && transb == Transpose::Trans)
     {
         args.set(3, one);
-        args.set(4, lda);
-        args.set(7, ldb);
+        args.set(4, (uint32_t)lda);
+        args.set(7, (uint32_t)ldb);
         args.set(8, one);
     }
     else
     {
         args.set(3, one);
-        args.set(4, lda);
+        args.set(4, (uint32_t)lda);
         args.set(7, one);
-        args.set(8, ldb);
+        args.set(8, (uint32_t)ldb);
     }
 
     KernelRange range(m, n);
